requirements.txt
# Python dependencies
pytest==7.4.3
python-dotenv==1.0.0
wordcloud==1.9.2
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.26.2
Pillow==10.1.0
